Move ConfigField and FieldFallback base classes back into confmodel.config.
