+++ docs/api_fallbacks.rst
@@ -1,0 +1,23 @@
+========================
+Field fallback reference
+========================
+
+
+:class:`~confmodel.fallbacks.FieldFallback` base class
+======================================================
+
+.. autoclass:: confmodel.fallbacks.FieldFallback
+   :members:
+
+
+:mod:`confmodel.fallbacks` module
+=================================
+
+All standard field fallback classes live here.
+
+TODO: Write docstrings for remaining more fallbacks.
+TODO: Document base class separately?
+
+.. automodule:: confmodel.fallbacks
+   :show-inheritance:
+   :members:
