+++ docs/api_fields.rst
@@ -1,0 +1,22 @@
+======================
+Config field reference
+======================
+
+
+:class:`~confmodel.fields.ConfigField` base class
+=================================================
+
+.. autoclass:: confmodel.fields.ConfigField
+   :members:
+
+
+:mod:`confmodel.fields` module
+==============================
+
+All standard config field classes live here.
+
+TODO: Write docstrings for remaining more fields.
+
+.. automodule:: confmodel.fields
+   :show-inheritance:
+   :members:
